Delete api/.env
The .env file is being deleted again since when the pull request was merged it created another one in the main. After this commit it is expected that when someone else pushes, the .env file will not be uploaded. If it happens again, it will be configured in a deeper and safer way using the command line in git bash.