Preview Page implemented
There are 4 buttons as placeholder to download, share, copy link and save. It is also as a placeholder, the place where the implementation of being able to view the PDF will be. All this together with its CSS attributes in a separate file. New dependency "@mui/icons-material". Command to install it "npm install @mui/icons-material